+++ services/vision/mlvision/classifier.go
@@ -3,10 +3,8 @@ package mlvision
 import (
 	"context"
 	"image"
-	"math"
 	"strconv"
 
-	"github.com/montanaflynn/stats"
 	"github.com/nfnt/resize"
 	"github.com/pkg/errors"
 
@@ -60,40 +58,4 @@ func attemptToBuildClassifier(mlm mlmodel.Service) (classification.Classifier, e
 		}
 		return classifications, nil
 	}, nil
-}
-
-// softmax takes the input slice and applies the softmax function.
-func softmax(in []float64) []float64 {
-	out := make([]float64, 0, len(in))
-	bigSum := 0.0
-	for _, x := range in {
-		bigSum += math.Exp(x)
-	}
-	for _, x := range in {
-		out = append(out, math.Exp(x)/bigSum)
-	}
-	return out
-}
-
-// checkClassification scores ensures that the input scores (output of classifier)
-// will represent confidence values (from 0-1).
-func checkClassificationScores(in []float64) []float64 {
-	if len(in) > 1 {
-		for _, p := range in {
-			if p < 0 || p > 1 { // is logit, needs softmax
-				confs := softmax(in)
-				return confs
-			}
-		}
-		return in // no need to softmax
-	}
-	// otherwise, this is a binary classifier
-	if in[0] < -1 || in[0] > 1 { // needs sigmoid
-		out, err := stats.Sigmoid(in)
-		if err != nil {
-			return in
-		}
-		return out
-	}
-	return in // no need to sigmoid
 }
